move auto hostendpoint logic to own file
